Laravel 5.2 displaying all posts and setting up

+++ app/Http/Controllers/AdminUsersController.php
@@ -132,10 +132,12 @@ class AdminUsersController extends Controller
     public function destroy($id)
     {
         //
-        $user = User::findOrFail($id);
-        unlink(public_path() . $user->photo->file);
-        $user -> delete();
 
+        $user = User::findOrFail($id);
+        if (!(empty($user->photo->file))){
+            unlink(public_path().  $user->photo->file);
+        }
+        $user -> delete();
         Session::flash('deleted_user', 'The user has been deleted');
